updated backend

+++ Smart_Accounting/Smart_Accounting.Application/Employee/Models/UpdatedEmployeeDto.cs
@@ -3,6 +3,6 @@ using System.ComponentModel.DataAnnotations;
 namespace Smart_Accounting.Application.Employee.Models {
     public class UpdatedEmployeeDto : EmployeDto {
         [Required]
-        public long id { get; set; }
+        public uint id { get; set; }
     }
 }
